requirements.txt
-f /usr/share/pip-wheels
Django==1.8.3
djangorestframework==3.1.3
ipdb==0.8.1
ipython==3.2.1
lxml==3.4.4
requests==2.7.0
wheel==0.24.0
whitenoise==2.0.2
